Cleaning up the python jwt code a bit.
* Moved jwt methods into own file
* Moved configuration parameters into the config file
* Removed unused jwt methods
